feat: add RatingOption component for heart, star, number variants
