+++ frontend/src/components/Field/Rating/Rating.tsx
@@ -1,121 +1,16 @@
-import {
-  Fragment,
-  HTMLProps,
-  KeyboardEvent,
-  useCallback,
-  useEffect,
-  useMemo,
-  useState,
-} from 'react'
+import { Fragment, useEffect, useMemo, useState } from 'react'
 import {
   Box,
   forwardRef,
+  Text,
   useBreakpointValue,
-  useRadio,
-  useToken,
   Wrap,
   WrapItem,
 } from '@chakra-ui/react'
 
 import { FieldColorScheme } from '~theme/foundations/colours'
 
-interface RatingComponent {
-  /**
-   * Radio styling props to spread on container.
-   */
-  checkbox: Omit<HTMLProps<any>, never>
-  /**
-   * Whether component is checked.
-   */
-  isChecked: boolean
-  /**
-   * ID of the input this component is rendering for.
-   */
-  inputId: string | undefined
-  /**
-   * Color scheme of the component to render. Defaults to `primary`.
-   */
-  colorScheme?: FieldColorScheme
-
-  children?: React.ReactNode
-}
-const NumberRating = ({
-  checkbox,
-  inputId,
-  isChecked,
-  colorScheme = 'primary',
-  children,
-}: RatingComponent): JSX.Element => {
-  const themeColorVar = useMemo(() => {
-    switch (colorScheme) {
-      case 'theme-red':
-      case 'theme-orange':
-      case 'theme-yellow':
-        return `${colorScheme}.700`
-      default:
-        return `${colorScheme}.500`
-    }
-  }, [colorScheme])
-
-  const themeColor = useToken('colors', themeColorVar)
-
-  return (
-    <Box
-      {...checkbox}
-      as="label"
-      htmlFor={inputId}
-      aria-hidden={false}
-      {...(isChecked ? { 'data-checked': '' } : {})}
-      minW="3.25rem"
-      display="flex"
-      justifyContent="center"
-      transitionProperty="common"
-      transitionDuration="normal"
-      cursor="pointer"
-      py="10px"
-      px="14px"
-      bg="white"
-      borderWidth="1px"
-      borderColor={themeColor}
-      color={themeColor}
-      _disabled={{
-        borderColor: 'neutral.500',
-        color: 'neutral.500',
-        cursor: 'not-allowed',
-
-        _checked: {
-          bg: 'neutral.500',
-          _hover: {
-            bg: 'neutral.500',
-          },
-        },
-        _hover: {
-          bg: 'white',
-        },
-        _active: {
-          color: 'neutral.500',
-          bg: 'white',
-        },
-      }}
-      _hover={{
-        bg: `${colorScheme}.200`,
-      }}
-      _active={{
-        bg: themeColor,
-        color: 'white',
-      }}
-      _focus={{
-        boxShadow: `0 0 0 4px var(--chakra-colors-${colorScheme}-300)`,
-      }}
-      _checked={{
-        bg: themeColor,
-        color: 'white',
-      }}
-    >
-      {children}
-    </Box>
-  )
-}
+import { RatingOption } from './RatingOption'
 
 export interface RatingProps {
   /**
@@ -150,110 +45,6 @@ export interface RatingProps {
    */
   variant: 'Heart' | 'Star' | 'Number'
 }
-
-interface RatingOptionProps {
-  /**
-   * Color scheme of the component to render. Defaults to `primary`.
-   */
-  colorScheme?: FieldColorScheme
-  /**
-   * Value of the option.
-   */
-  value: number
-  /**
-   * Function called once a rating is selected.
-   * @param newRating the value of the checked radio
-   */
-  onChange?: (newRating: number | undefined) => void
-  /**
-   * The current selected value in the rating group.
-   */
-  selectedValue?: number
-  /**
-   * The `name` attribute forwarded to the rating `radio` element
-   */
-  name: string
-
-  /**
-   * Variant of rating field to render
-   */
-  variant: 'Heart' | 'Star' | 'Number'
-}
-
-const RatingOption = forwardRef<RatingOptionProps, 'input'>(
-  (
-    {
-      children,
-      colorScheme = 'primary',
-      name,
-      onChange,
-      selectedValue,
-      value,
-      variant,
-    },
-    ref,
-  ) => {
-    const handleSelect = useCallback(() => {
-      if (selectedValue === value) {
-        onChange?.(undefined)
-      } else {
-        onChange?.(value)
-      }
-    }, [onChange, selectedValue, value])
-
-    const handleSpacebar = useCallback(
-      (e: KeyboardEvent<HTMLInputElement>) => {
-        if (e.key !== ' ') return
-        if (selectedValue === value) {
-          e.preventDefault()
-          handleSelect()
-        }
-      },
-      [handleSelect, selectedValue, value],
-    )
-
-    const { getInputProps, getCheckboxProps } = useRadio({
-      name,
-      id: `${name}-${value}`,
-      onChange: handleSelect,
-      value,
-    })
-
-    const input = getInputProps()
-    const checkbox = getCheckboxProps()
-
-    const ComponentToRender = useMemo(() => {
-      switch (variant) {
-        case 'Number':
-        case 'Heart':
-        case 'Star':
-        default:
-          return NumberRating
-      }
-    }, [variant])
-
-    return (
-      <Box _active={{ zIndex: 1 }} _focusWithin={{ zIndex: 1 }}>
-        <input
-          type="radio"
-          aria-checked={selectedValue === value}
-          {...input}
-          onChange={handleSelect}
-          onKeyDown={handleSpacebar}
-          ref={ref}
-        />
-        <ComponentToRender
-          checkbox={checkbox}
-          isChecked={selectedValue === value}
-          colorScheme={colorScheme}
-          inputId={input.id}
-        >
-          {children}
-        </ComponentToRender>
-      </Box>
-    )
-  },
-)
 
 export const Rating = forwardRef<RatingProps, 'input'>(
   (
@@ -293,35 +84,54 @@ export const Rating = forwardRef<RatingProps, 'input'>(
       [numberOfRatings],
     )
 
+    const ratingLayout = useMemo(() => {
+      switch (variant) {
+        case 'Number':
+          return { spacing: '-px', rowHeight: '0.5rem' }
+        default:
+          return { spacing: '0.25rem', rowHeight: 0 }
+      }
+    }, [variant])
+
     return (
-      <Wrap as="fieldset" spacing="-px">
-        {options.map((value, i) => {
-          return (
-            <Fragment key={value}>
-              <WrapItem>
-                <RatingOption
-                  name={name}
-                  variant={variant}
-                  colorScheme={colorScheme}
-                  value={value}
-                  numberOfRatings={numberOfRatings}
-                  onChange={setCurrentValue}
-                  selectedValue={currentValue}
-                  // Pass in ref if first item so it can be focused.
-                  {...(i === 0 ? { ref } : {})}
-                >
-                  {value}
-                </RatingOption>
-              </WrapItem>
-              {
-                // Force component to begin on a new line.
-                value % wrapComponentsPerRow === 0 && isSplitRows && (
-                  <Box flexBasis="100%" h="0.5rem" />
-                )
-              }
-            </Fragment>
-          )
-        })}
+      <Wrap
+        as="fieldset"
+        align="center"
+        shouldWrapChildren
+        spacing={['0.5rem', '0.5rem', '1rem']}
+      >
+        <Wrap spacing={ratingLayout.spacing}>
+          {options.map((value, i) => {
+            return (
+              <Fragment key={value}>
+                <WrapItem>
+                  <RatingOption
+                    name={name}
+                    variant={variant}
+                    colorScheme={colorScheme}
+                    value={value}
+                    numberOfRatings={numberOfRatings}
+                    onChange={setCurrentValue}
+                    selectedValue={currentValue}
+                    // Pass in ref if first item so it can be focused.
+                    {...(i === 0 ? { ref } : {})}
+                  >
+                    {value}
+                  </RatingOption>
+                </WrapItem>
+                {
+                  // Force component to begin on a new line.
+                  value % wrapComponentsPerRow === 0 && isSplitRows && (
+                    <Box flexBasis="100%" h={ratingLayout.rowHeight} />
+                  )
+                }
+              </Fragment>
+            )
+          })}
+        </Wrap>
+        {currentValue && variant !== 'Number' && (
+          <Text textStyle="subhead-2">{currentValue} selected</Text>
+        )}
       </Wrap>
     )
   },
