+++ frontend/src/components/Field/Rating/index.ts
@@ -1,0 +1,1 @@
+export { Rating as default } from './Rating'
